--- a/js/data/bucket.js
+++ b/js/data/bucket.js
@@ -71,7 +71,8 @@ function Bucket(options) {
     this.type = this.layer.type;
     this.features = [];
     this.id = this.layer.id;
-    this['source-layer'] = this.layer['source-layer'];
+    this.sourceLayer = this.layer.sourceLayer;
+    this.sourceLayerIndex = options.sourceLayerIndex;
     this.minZoom = this.layer.minzoom;
     this.maxZoom = this.layer.maxzoom;