This builds on https://github.com/mapbox/mapbox-gl-js/pull/2052.

This is not ready, but it could use some high level :eyes:. I probably should have gotten some :eyes: a bit earlier. Let me know if any of the high level ideas here concern you. I'll comment again later when this is more polished.

The raw vector tile pbf, FeatureTree and CollisionTile are now transferred to the main thread when a tile finishes loading. This lets use have a synchronous `queryRenderedFeatures`, it lets us eliminate `includeGeometry` without a performance hit, and it starts moving us towards more state-less workers.

### A synchronous queryRenderedFeatures
This adds a synchronous version of `queryRenderedFeatures`.

- it's faster
- it's more convenient to use, especially when you want to make multiple queries
- it's more consistent. It never gets delayed for 100s of ms while a worker is busy parsing a tile
- you don't have to worry about callbacks being called out of order
- things like https://github.com/mapbox/mapbox-gl-js/issues/1002 could be done lazily

There is still an async version. It transfers data back to workers.

Despite being parallelized, the async version is not faster than the sync version. It ends up parsing from the pbfs twice instead of once. It spends less time on the main thread though which could be useful for large, expensive queries. I think it could be worthwhile to keep both versions as `queryRenderedFeatures` and `queryRenderedFeaturesAsync`

### Getting rid of `includeGeometry`

The `includeGeometry` param for `queryRenderedFeatures` only exists because sending geometries from the worker back to the main thread was slow. When the pbf is on the main thread individual geometries can be deserialized **lazily**. This lets us drop `includeGeometry` without a performance penalty.

### indexing with Grid

I replaced rbush with a [`Grid`](https://github.com/mapbox/mapbox-gl-js/blob/query-rendered-features-continued/js/util/grid.js), a different kind of spatial index. `Grid` can be serialized to an ArrayBuffer to make it transferrable. The serialized version can be queried directly without deserializing individual elements to javascript objects. It is [faster than rbush for collision detection](https://github.com/mapbox/mapbox-gl-js/commit/0ec60ff96749276f32710ec43f26ccff03f50096) and comparable for [rendered feature querying](https://github.com/mapbox/mapbox-gl-js/commit/6e3f8c7a2a2388718afa5c2dc4d64395a6c82f5d). It's faster because insertions are faster. 

 Grid is a grid of cells. When a bbox is inserted it is added to all the cells it intersects. When a bbox is queried it looks in all cells that intersects with that bbox. That's it.

It isn't faster than rbush generally, but it is faster for what we use it for: a fixed extent, mostly evenly sized bboxes, and a high number of insertions. More importantly, it's transferable. 

### Storing arrays of "structs" in ArrayBuffers
There are a bunch of cases where we have giant arrays of fixed size objects. I added a new abstraction, [StructArray](https://github.com/mapbox/mapbox-gl-js/blob/query-rendered-features-continued/js/util/struct_array.js), that makes it easier to store these kinds of objects in ArrayBuffers. It's inspired by the new Buffer implementation. It:
- saves a lot of space. Switching collision boxes saved 500KB per tile.
- makes transferring or even copying these arrays a lot faster
- avoids creating a ton of garbage

see https://github.com/mapbox/mapbox-gl-js/commit/9a2e5fd4645cde999005796538da975ab5b5d217 's commit message for an example

### Reducing worker state
Both Grid and StructArrays are transferrable and don't need to be deserialized to individual objects before they can be used. This makes moving them between workers and the main thread really fast and reduces the amount of data that is "stuck" on workers. When `symbolInstances` is made completely transferable we could get rid of WorkerTiles that live after parsing has finished.

The work requests would transfer all the data that is needed and not rely on the worker having any unique data. I think this is also more similar to how we want work requests to work in -native. 

Since work won't be tied to a specific thread, it should be possible to distribute work more consistently and speed things up.

### Benchmarks

Compared to https://github.com/mapbox/mapbox-gl-js/pull/2052

```

point query
http://localhost:9966/debug/#15.37/38.9047/-77.0299
mapbox-streets-v6 (before all features were merged)
new sync                        1 ms
new sync with geometry          1 ms
new async                       5-6 ms
new async with geometry         5-6 ms
old async                       1-2 ms
old async with geometry         2 ms

point query
http://localhost:9966/debug/#15.37/38.9047/-77.0299
mapbox-streets-v7 (after many layers are merged into a single huge multi-geometry feature)
new sync                        2 ms
new sync with geometry          2 ms
new async                       5 ms
new async with geometry         5-6 ms
old async                       5 ms
old async with geometry         15-30 ms

full screen box query
http://localhost:9966/debug/#15.37/38.9047/-77.0299
mapbox-streets-v7 (after many layers are merged into a single huge multi-geometry feature)
new sync                        23 ms
new sync with geometry          70 ms
new async                       19-25 ms
new async with geometry         75 ms
old async                       45 ms
old async with geometry         770 ms


full screen bbox query
http://localhost:9966/debug/#11.52/38.9262/-77.0420
mapbox-streets-v7 (after many layers are merged into a single huge multi-geometry feature)
new sync                        115 ms
new sync with geometry          175 ms
new aysnc                       125 ms
new async with geometry         180 ms
old async                       265 ms
old async with geometry         560 ms
```

- The old and new sync queryRenderedFeatures are both really fast for point queries.
- The new queryRenderedFeatures is much faster for large box queries, especially if geometries are included.
- The new sync and async queryRenderedFeatures take a similar amount of time, but the async version spends less time on the main thread.

### next steps
Cleaning things up. Fixing tests. Fixing a couple remaining bugs.

Do all the high level changes sound ok?

@lucaswoj @jfirebaugh @mourner