@mourner 2016-03-07
What's the memory impact of switching to grid index and storing it in both the main thread and the worker?