@lucaswoj 2016-03-07
I'm very :+1: on the high-level changes in this PR.

Before shipping, I'd like for us to think about 

 - what it'd take to merge `StructArray` and `Buffer` ([LOC is proportional to # of bugs](http://www.techrepublic.com/blog/it-security/the-danger-of-complexity-more-code-more-bugs/))
 - how stateless workers can be used to write better integration tests for mapbox-gl-js (the tests can be another PR but lets think about the architecture that enables those tests now)
 - how we can codify the performance benchmarks used in this PR into something reusable