@ansis 2016-03-08
>  - what it'd take to merge `StructArray` and `Buffer` ([LOC is proportional to # of bugs](http://www.techrepublic.com/blog/it-security/the-danger-of-complexity-more-code-more-bugs/))

Buffer does a whole bunch of extra calculations like scaling values and it doesn't provide an interface for reading values. I think it might make sense to wait until dds settles down before trying to merge the two.

>  - how stateless workers can be used to write better integration tests for mapbox-gl-js (the tests can be another PR but lets think about the architecture that enables those tests now)

How would stateless workers help us write better integration tests?

> - how we can codify the performance benchmarks used in this PR into something reusable 

Yes, I definitely want to write reusable benchmarks. Do we want to start using a framework for this or continue writing it ourselves (like with the buffer benchmark)?