@lucaswoj 2016-03-08
> Buffer does a whole bunch of extra calculations like scaling values 

That's not correct. Those "extra calculations" are done in `Bucket`. Buffer is designed to be a lightweight generic `StructArray` implementation. 

> and it doesn't provide an interface for reading values. 

It has a [`Buffer#get` method](https://github.com/mapbox/mapbox-gl-js/blob/master/js/data/buffer.js#L152) for reading values. The docs say that it is "only used for debugging" as an indication that it is dead code, not as an indication of any technical limitations. 

> I think it might make sense to wait until dds settles down before trying to merge the two.

The `Buffer` class will not change in any future dds work.

-----

> How would stateless workers help us write better integration tests?

Out of scope for this PR. I'm excited about the implications of a simpler interface for writing better integration tests through `skipWorker`-like functionality #1979.

`test-suite` is slow and limited in what it can test. https://github.com/mapbox/mapbox-gl-js/issues/2223 is a perfect example of a bug that we won't be able to write regression tests for and maybe would've been caught if we had better integration tests.

----- 

> Yes, I definitely want to write reusable benchmarks. Do we want to start using a framework for this or continue writing it ourselves (like with the buffer benchmark)?

I'm going to design a simple benchmark interface specification today. It'll be similar to how the buffer benchmark works.