@lucaswoj 2016-03-08
benchmark interface specification pushed to #2231 :smile: