@ansis 2016-03-09
Rebased on master. Latest commit used to be 3fc75bd3127436715d0ec2e88d52aade617ea526.