@lucaswoj 2016-03-15
A convention I've been using, and have found fruitful:

 - object constructors take a single `options` argument 
 - `serialize` methods output an object compatible with the constructor `options` argument