@ansis 2016-03-16
> A convention I've been using, and have found fruitful
> ...

switched in fe8dad80fb4ce20d357b01bdc899a230289e09f6

----

and then I squashed all those commits that fixed things from d40a059ef0f94b10a93f0578658213ac5a21229