@mcwhittemore 2016-03-17
Any time frame here? This would solve a bunch of event handler problems for gl-draw.