@ansis 2016-03-17
@mcwhittemore soon, hopefully

@lucaswoj I added migration instructions to changelog, a benchmark for point queries, a benchmark for full-screen box queries, and updated all examples. All that's left is review and fixing things based on that.